requirement.txt
dash=2.6.0
pandas
gunicorn
plotly.express